+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - create location only once and save as reported property
 - ignored RUSTSEC-2020-0071
 - removed unused dependency to notify
+- bumped to latest azure-iot-sdk 0.8.2
 
 ## [0.1.0] Q3 2022
 - initial version